+++ addon/services/portal-service.js
@@ -165,9 +165,17 @@ export default Service.extend(serviceMixin, {
     return this.request(urlPath, null, portalOpts);
   },
 
+  /**
+   * Config object defining ability to auto-join org with social logins
+   */
+  getSocialProviders (portalOpts) {
+    const urlPath = `/portals/self/socialProviders?f=json`;
+    return this.request(urlPath, null, portalOpts);
+  },
+
   configureSocialProviders (opts, portalOpts) {
     /*
-      portalOpts: {
+      opts: {
         signUpMode: Automatic | ???
         providers: facebook,google
         level: 1 | 2
@@ -178,6 +186,14 @@ export default Service.extend(serviceMixin, {
     */
     const urlPath = `/portals/self/socialProviders/configure?f=json`;
     return this._post(urlPath, opts, portalOpts);
+  },
+
+  /**
+   * Remove a Social Providers
+   */
+  removeSocialProviders (portalOpts) {
+    const urlPath = `/portals/self/socialProviders/remove?f=json`;
+    return this.request(urlPath, { method: 'POST' }, portalOpts);
   },
 
   checkUsername (username, portalOpts) {
@@ -277,6 +293,14 @@ export default Service.extend(serviceMixin, {
   getSigninSettings (portalOpts) {
     const urlPath = `/portals/self/signinSettings?f=json`;
     return this.request(urlPath, {}, portalOpts);
+  },
+
+  updateContacts (users = [], portalOpts) {
+    /*
+      users: [user1Username, user2Username]
+    */
+    const urlPath = `/portals/self/update?contacts=${users}&f=json`;
+    return this._post(urlPath, null, portalOpts);
   }
 
 });